+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 
+## [0.8.0] - 2022-08-04
+
+### Added
+- `ZLayer.requires`
+- `ZLayer.identity`
+- `ZLayer.>>>`
+- `ZLayer.++`
+- `ZIO.provideSomeLayer`
+- `ZEnv.any`
+- `console.Console.any`
+
+### Changed
+- Program uses new DSL
+
+
 ## [0.7.0] - 2022-08-03
 
 ### Added
@@ -80,7 +95,8 @@ and this project adheres to [Semantic Versioning].
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[Unreleased]: https://github.com/jaacko-torus/diy-zio/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/jaacko-torus/diy-zio/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/jaacko-torus/diy-zio/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/jaacko-torus/diy-zio/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/jaacko-torus/diy-zio/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/jaacko-torus/diy-zio/compare/v0.4.0...v0.5.0